move string typing functions to new lib module. Instead of directly assigning trimatix.botState.client, use a package init method
