+++ src/trimatix/__init__.py
@@ -1,0 +1,7 @@
+from discord import Client
+
+def init(client: Client):
+    if not isinstance(client, Client):
+        raise TypeError("Expected type discord.Client, received " + type(client).__name__)
+    from . import botState
+    botState.client = client
