+++ src/trimatix/lib/stringTyping.py
@@ -1,0 +1,25 @@
+def strIsInt(x) -> bool:
+    """Decide whether or not something is either an integer, or is castable to integer.
+
+    :param x: The object to type-check
+    :return: True if x is an integer or if x can be casted to integer. False otherwise
+    :rtype: bool
+    """
+
+    try:
+        int(x)
+    except TypeError:
+        return False
+    except ValueError:
+        return False
+    return True
+
+
+def strIsRoleMention(mention: str) -> bool:
+    """Decide whether the given string is a discord role mention, being <@&ROLEID> where ROLEID is an integer discord role id.
+
+    :param str mention: The string to check
+    :return: True if mention matches the formatting of a discord role mention, False otherwise
+    :rtype: bool
+    """
+    return mention.endswith(">") and mention.startswith("<@&") and strIsInt(mention[3:-1])
